test(vitest): keep command fixture branches out of tests
